rbnspots-clubs.pl: Reload *full* $callhash with member info on member list change
Previously, we only reloaded member list of clubs that changed. This
works fine for *adding* new members but if a member disappears from
a list, it won't be removed from %callhash. So the only way to achieve
this is to completely clear the %callhash and rebuild it from all member
lists.

(tnx F5IYJ for finding and reporting this problem!)